% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_rdd.R
\name{approx_count}
\alias{approx_count}
\title{Find the approximate total number of records within a Spatial RDD.}
\usage{
approx_count(x)
}
\arguments{
\item{x}{A Sedona spatial RDD.}
}
\description{
Given a Sedona spatial RDD, find the (possibly approximated) number of total
records within it.
}
\seealso{
Other Spatial RDD aggregation routine: 
\code{\link{minimum_bounding_box}()}
}
\concept{Spatial RDD aggregation routine}
