% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{as.spark.dataframe}
\alias{as.spark.dataframe}
\title{Import data from a spatial RDD into a Spark Dataframe.}
\usage{
as.spark.dataframe(x, non_spatial_cols = NULL, name = NULL)
}
\arguments{
\item{x}{A spatial RDD.}

\item{non_spatial_cols}{Column names for non-spatial attributes in the
resulting Spark Dataframe.}

\item{name}{Name to assign to the resulting Spark temporary view. If
unspecified, then a random name will be assigned.}
}
\description{
Import data from a spatial RDD (possibly with non-spatial attributes) into a
Spark Dataframe.
}
