% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{as_spark_dataframe}
\alias{as_spark_dataframe}
\title{Import data from a spatial RDD into a Spark Dataframe.}
\arguments{
\item{x}{A spatial RDD.}

\item{name}{Name to assign to the resulting Spark temporary view. If
unspecified, then a random name will be assigned.}
}
\description{
Import data from a spatial RDD into a Spark Dataframe.
}
