% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounding_box.R
\name{bounding_box}
\alias{bounding_box}
\title{Find axis-aligned minimum bounding box of a geometry.}
\usage{
bounding_box(x)
}
\arguments{
\item{x}{A spatial RDD.}
}
\description{
Given a spatial RDD, find an axis-aligned minimum bounding box (envelope)
containing the geometrical object(s) represented by the RDD.
}
