% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_rdd.R
\name{crs_transform}
\alias{crs_transform}
\title{Perform a CRS transformation.}
\usage{
crs_transform(x, src_epsg_crs_code, dst_epsg_crs_code, strict = FALSE)
}
\arguments{
\item{x}{The spatial RDD to be processed.}

\item{src_epsg_crs_code}{Coordinate reference system to transform from.}

\item{dst_epsg_crs_code}{Coordinate reference system to transform to.}

\item{strict}{If FALSE (default), then ignore the "Bursa-Wolf Parameters
Required" error.}
}
\description{
Transform data within a spatial RDD from one coordinate reference system to
another.
}
