% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_rdd.R
\name{minimum_bounding_box}
\alias{minimum_bounding_box}
\title{Find the minimal bounding box of a geometry.}
\usage{
minimum_bounding_box(x)
}
\arguments{
\item{x}{A Sedona spatial RDD.}
}
\description{
Given a Sedona spatial RDD, find the axis-aligned minimal bounding box of the
geometry represented by the RDD.
}
\seealso{
Other Spatial RDD aggregation routine: 
\code{\link{approx_count}()}
}
\concept{Spatial RDD aggregation routine}
