% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_interface.R
\name{sdf_register.spatial_rdd}
\alias{sdf_register.spatial_rdd}
\title{Import data from a spatial RDD into a Spark Dataframe.}
\usage{
\method{sdf_register}{spatial_rdd}(x, name = NULL)
}
\arguments{
\item{x}{A spatial RDD.}
}
\description{
Import data from a spatial RDD into a Spark Dataframe.
}
\examples{
library(sparklyr)
library(sparklyr.sedona)

sc <- spark_connect(master = "spark://HOST:PORT")

if (!inherits(sc, "test_connection")) {
  input_location <- system.file(
    file.path("extdata", "polygon.json"), package = "sparklyr.sedona"
  )
  rdd <- sedona_read_geojson_to_typed_rdd(
    sc,
    location = input_location,
    type = "polygon"
  )
  sdf <- sdf_register(rdd)
}

}
