% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_queries.R
\name{sedona_range_query}
\alias{sedona_range_query}
\title{Execute a range query.}
\usage{
sedona_range_query(
  rdd,
  x,
  query_type = c("cover", "intersect"),
  index_type = c("quadtree", "rtree"),
  result_type = c("rdd", "sdf", "raw")
)
}
\arguments{
\item{rdd}{A Sedona spatial RDD.}

\item{x}{The query object.}

\item{query_type}{Type of spatial relationship involved in the query.
Currently "cover" and "intersect" are supported.}

\item{index_type}{Index to use to facilitate the KNN query. If NULL, then
do not build any additional spatial index on top of \code{x}. Supported
index types are "quadtree" and "rtree".}

\item{result_type}{Type of result to return.
If "rdd" (default), then the k nearest objects will be returned in a Sedona
spatial RDD.
If "sdf", then a Spark dataframe containing the k nearest objects will be
returned.
If "raw", then a list of k nearest objects will be returned. Each element
within this list will be a JVM object of type
\code{org.locationtech.jts.geom.Geometry}.}
}
\description{
Given a spatial RDD and a query object \code{x}, find all spatial objects
within the RDD that are covered by \code{x} or intersect \code{x}.
}
\seealso{
Other Sedona spatial query: 
\code{\link{sedona_knn_query}()}
}
\concept{Sedona spatial query}
