% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{sedona_read_dsv}
\alias{sedona_read_dsv}
\title{Create a typed SpatialRDD from a delimiter-separated values file.}
\usage{
sedona_read_dsv(
  sc,
  location,
  delimiter = c(",", "\\t", "?", "'", "\\"", "_", "-", "\%", "~", "|", ";"),
  type = c("point", "polygon", "linestring"),
  first_spatial_col_index = 0L,
  last_spatial_col_index = NULL,
  contains_non_geom_attributes = TRUE,
  memory = TRUE,
  repartition = 1L
)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{location}{Location of the input file.}

\item{delimiter}{Delimiter within each record. Must be one of
',', '\\t', '?', '\\'', '"', '_', '-', '\%', '~', '|', ';'}

\item{type}{Type of the SpatialRDD (must be one of "point", "polygon", or
"linestring".}

\item{first_spatial_col_index}{Zero-based index of the left-most column
containing spatial attributes (default: 0).}

\item{last_spatial_col_index}{Zero-based index of the right-most column
containing spatial attributes (default: NULL). Note last_spatial_col_index
does not need to be specified when creating a PointRDD because it will
automatically have the implied value of (first_spatial_col_index + 1).
For all other types of RDDs, if last_spatial_col_index is unspecified, then
it will assume the value of -1 (i.e., the last of all input columns).}

\item{contains_non_geom_attributes}{Whether the input contains non-
geometrical attributes (default: TRUE).}

\item{memory}{Whether to cache the resulting RDD in memory (default: TRUE).}

\item{repartition}{The number of partitions to have in the resulting RDD
(default: 1).}
}
\description{
Create a typed SpatialRDD (namely, a PointRDD, a PolygonRDD, or a
LineStringRDD) from a delimiter-separated values file.
The input file can contain spatial attributes (e.g., longitude and latidude)
and other attributes. Currently only inputs with spatial attributes occupying
a contiguous range of columns (i.e.,
[first_spatial_col_index, last_spatial_col_index]) are supported.
}
