% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{sedona_read_shapefile_to_typed_rdd}
\alias{sedona_read_shapefile_to_typed_rdd}
\title{Create a typed SpatialRDD from a shapefile data source.}
\usage{
sedona_read_shapefile_to_typed_rdd(
  sc,
  location,
  type = c("point", "polygon", "linestring"),
  storage_level = "MEMORY_ONLY"
)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{location}{Location of the data source.}

\item{type}{Type of the SpatialRDD (must be one of "point", "polygon", or
"linestring".}

\item{storage_level}{Storage level of the RDD (default: MEMORY_ONLY).}
}
\description{
Create a typed SpatialRDD (namely, a PointRDD, a PolygonRDD, or a
LineStringRDD) from a shapefile data source.
}
\examples{
library(sparklyr)
library(sparklyr.sedona)

sc <- spark_connect(master = "spark://HOST:PORT")

if (!inherits(sc, "test_connection")) {
  input_location <- system.file(
    file.path("extdata", "shapefile-example"), package = "sparklyr.sedona"
  )
  rdd <- sedona_read_shapefile_to_typed_rdd(
    sc, location = input_location, type = "polygon"
  )
}

}
