% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{sedona_read_typed_shapefile}
\alias{sedona_read_typed_shapefile}
\title{Create a typed SpatialRDD from a shapefile data source.}
\usage{
sedona_read_typed_shapefile(
  sc,
  location,
  type = c("point", "polygon", "linestring"),
  storage_level = "MEMORY_ONLY"
)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{location}{Location of the data source.}

\item{type}{Type of the SpatialRDD (must be one of "point", "polygon", or
"linestring".}

\item{storage_level}{Storage level of the RDD (default: MEMORY_ONLY).}
}
\description{
Create a typed SpatialRDD (namely, a PointRDD, a PolygonRDD, or a
LineStringRDD) from a shapefile data source.
}
