% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{sedona_spatial_rdd_serialization_routine}
\alias{sedona_spatial_rdd_serialization_routine}
\title{Write SpatialRDD into a file.}
\arguments{
\item{x}{The SpatialRDD object.}

\item{output_location}{Location of the output file.}
}
\description{
Export serialized data from a Sedona SpatialRDD into an output file.
}
