% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{sedona_write_wkb}
\alias{sedona_write_wkb}
\title{Write SpatialRDD into a WKB file.}
\usage{
sedona_write_wkb(x, output_location)
}
\arguments{
\item{x}{The SpatialRDD object.}

\item{output_location}{Location of the output file.}
}
\description{
Export serialized data from a Sedona SpatialRDD into a WKB file.
}
\examples{
library(sparklyr)
library(sparklyr.sedona)

sc <- spark_connect(master = "spark://HOST:PORT")

if (!inherits(sc, "test_connection")) {
  input_location <- system.file(
    file.path("extdata", "county_tiny_wkb.tsv"), package = "sparklyr.sedona"
  )
  rdd <- sedona_read_wkb(
    sc,
    location = input_location,
    wkb_col_idx = 0L
  )
  sedona_write_wkb(rdd, "/tmp/wkb_output.tsv")
}

}
\seealso{
Other Sedona data inferface functions: 
\code{\link{sedona_read_dsv_to_typed_rdd}()},
\code{\link{sedona_read_geojson_to_typed_rdd}()},
\code{\link{sedona_read_geojson}()},
\code{\link{sedona_read_shapefile}()},
\code{\link{sedona_read_wkb}()},
\code{\link{sedona_read_wkt}()},
\code{\link{sedona_save_spatial_rdd}()},
\code{\link{sedona_write_geojson}()},
\code{\link{sedona_write_wkt}()}
}
\concept{Sedona data inferface functions}
