% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interface.R
\name{sedona_read_geojson}
\alias{sedona_read_geojson}
\title{Create a typed SpatialRDD from a GeoJSON data source.}
\usage{
sedona_read_geojson(
  sc,
  location,
  type = c("point", "polygon", "linestring"),
  contains_non_geom_attributes = TRUE,
  storage_level = "MEMORY_ONLY",
  repartition = 1L
)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{location}{Location of the data source.}

\item{type}{Type of the SpatialRDD (must be one of "point", "polygon", or
"linestring".}

\item{contains_non_geom_attributes}{Whether the input contains non-
geometrical attributes (default: TRUE).}

\item{storage_level}{Storage level of the RDD (default: MEMORY_ONLY).}

\item{repartition}{The number of partitions to have in the resulting RDD
(default: 1).}
}
\description{
Create a typed SpatialRDD (namely, a PointRDD, a PolygonRDD, or a
LineStringRDD) from a GeoJSON data source.
}
